refactor: counter example move to content folder
