+++ src/app/store.ts
@@ -13,7 +13,7 @@ import {
 } from 'reduxjs-toolkit-persist';
 import { WebStorage } from 'reduxjs-toolkit-persist/lib/types';
 
-import counterReducer from '../features/counter/counterSlice';
+import counterReducer from '../content/features/counter/counterSlice';
 
 const persistConfig = {
   key: 'root',
