+++ src/content/features/counter/Counter.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 
-import { useAppDispatch, useAppSelector } from '../../app/hooks';
+import { useAppDispatch, useAppSelector } from '../../../app/hooks';
 
 import {
   decrement,
