+++ src/content/features/counter/counterAPI.ts
