+++ src/content/features/counter/counterSlice.ts
@@ -1,6 +1,6 @@
 import { createAsyncThunk, createSlice, PayloadAction } from '@reduxjs/toolkit';
 
-import { AppThunk, RootState } from '../../app/store';
+import { AppThunk, RootState } from '../../../app/store';
 
 import { fetchCount } from './counterAPI';
