+++ src/content/features/counter/index.ts
@@ -1,0 +1,2 @@
+export { Counter } from './Counter';
+export { default as counterReducer } from './counterSlice';
